source structure, move binding module public headers
DO NOT MERGE

Move binding public header files to include/bluetoe. This is a follow-up
to the previous commit.
